requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7
psycopg2-binary==2.9.7
cloud-sql-python-connector[pg8000]==1.4.0
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
google-generativeai==0.3.0
pydub==0.25.1
requests==2.31.0
elevenlabs==0.2.26
pytz==2023.3
python-dateutil==2.8.2
